# apps/web/lib/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MedBaza API",
    "description": "Multi-vendor marketplace for medical supplies and equipment. Money is always integer minor units plus an ISO-4217 currency.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/checkout": {
      "post": {
        "tags": [
          "checkout"
        ],
        "summary": "Create Checkout",
        "description": "Place an order for everything in the cart.\n\nRequires an `Idempotency-Key`: a repeated key returns the original order\nrather than creating a second one. Totals and stock are decided by the\nserver; anything the client sent about price is ignored (§5.2, §5.6).",
        "operationId": "createCheckout",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Cart-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque guest-cart token. Ignored once the caller is signed in.",
              "title": "X-Cart-Token"
            },
            "description": "Opaque guest-cart token. Ignored once the caller is signed in."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client-generated key; retries reuse it.",
              "title": "Idempotency-Key"
            },
            "description": "Client-generated key; retries reuse it."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressInput": {
        "properties": {
          "recipient_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Recipient Name"
          },
          "line1": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Line1"
          },
          "line2": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Line2"
          },
          "city": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "City"
          },
          "region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "postal_code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Postal Code"
          },
          "country": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Country"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "recipient_name",
          "line1",
          "city",
          "postal_code",
          "country"
        ],
        "title": "AddressInput"
      },
...
      "CheckoutRequest": {
        "properties": {
          "shipping_address": {
            "$ref": "#/components/schemas/AddressInput"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "prescription_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prescription Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "shipping_address"
        ],
        "title": "CheckoutRequest"
      },
      "CheckoutResponse": {
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderRead"
          },
          "payment_redirect_url": {
            "type": "string",
            "title": "Payment Redirect Url"
          }
        },
        "type": "object",
        "required": [
          "order",
          "payment_redirect_url"
        ],
        "title": "CheckoutResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderEventRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "shipment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipment Id"
          },
          "from_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Status"
          },
          "to_status": {
            "type": "string",
            "title": "To Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "shipment_id",
          "from_status",
          "to_status",
          "reason",
          "created_at"
        ],
        "title": "OrderEventRead"
      },
      "OrderItemRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "product_slug": {
            "type": "string",
            "title": "Product Slug"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_amount_minor": {
            "type": "integer",
            "title": "Unit Amount Minor"
          },
          "line_amount_minor": {
            "type": "integer",
            "title": "Line Amount Minor"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "prescription_required": {
            "type": "boolean",
            "title": "Prescription Required"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "product_name",
          "product_slug",
          "sku",
          "image_url",
          "quantity",
          "unit_amount_minor",
          "line_amount_minor",
          "currency",
          "prescription_required"
        ],
        "title": "OrderItemRead",
        "description": "Rendered entirely from the purchase-time snapshot (§5.2)."
      },
...
      "OrderRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "items_amount_minor": {
            "type": "integer",
            "title": "Items Amount Minor"
          },
          "shipping_amount_minor": {
            "type": "integer",
            "title": "Shipping Amount Minor"
          },
          "tax_amount_minor": {
            "type": "integer",
            "title": "Tax Amount Minor"
          },
          "total_amount_minor": {
            "type": "integer",
            "title": "Total Amount Minor"
          },
          "shipping_address": {
            "$ref": "#/components/schemas/AddressInput"
          },
          "prescription_required": {
            "type": "boolean",
            "title": "Prescription Required"
          },
          "prescription_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prescription Id"
          },
          "prescription_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prescription Status"
          },
          "shipments": {
            "items": {
              "$ref": "#/components/schemas/ShipmentRead"
            },
            "type": "array",
            "title": "Shipments"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/OrderEventRead"
            },
            "type": "array",
            "title": "Events"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "status",
          "currency",
          "items_amount_minor",
          "shipping_amount_minor",
          "tax_amount_minor",
          "total_amount_minor",
          "shipping_address",
          "prescription_required",
          "prescription_id",
          "prescription_status",
          "shipments",
          "events",
          "created_at"
        ],
        "title": "OrderRead"
      },
...
      "ShipmentRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "seller_id": {
            "type": "string",
            "title": "Seller Id"
          },
          "seller_name": {
            "type": "string",
            "title": "Seller Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "carrier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carrier"
          },
          "tracking_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Number"
          },
          "items_amount_minor": {
            "type": "integer",
            "title": "Items Amount Minor"
          },
          "shipping_amount_minor": {
            "type": "integer",
            "title": "Shipping Amount Minor"
          },
          "total_amount_minor": {
            "type": "integer",
            "title": "Total Amount Minor"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "seller_id",
          "seller_name",
          "status",
          "carrier",
          "tracking_number",
          "items_amount_minor",
          "shipping_amount_minor",
          "total_amount_minor",
          "currency",
          "items"
        ],
        "title": "ShipmentRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}